src/requirements.txt
termcolor
colorama
antlr4-python3-runtime==4.7.1
autopep8==1.3.5
py-mini-racer==0.1.13
pycodestyle==2.4.0
